+++ src/components/timeline/Row.svelte
@@ -53,6 +53,7 @@
   import { getDoyTime } from '../../utilities/time';
   import {
     directiveInView,
+    generateActivityTree as generateActivityTreeUtil,
     getYAxesWithScaleDomains,
     isActivityLayer,
     isLineLayer,
@@ -418,83 +419,24 @@
     timeFilteredActivityDirectives &&
     timeFilteredSpans &&
     activityOptions &&
+    activityTreeExpansionMap &&
     typeof showSpans === 'boolean' &&
-    typeof showDirectives === 'boolean' &&
-    activityTreeExpansionMap
+    typeof showDirectives === 'boolean'
   ) {
-    // TODO figure out how to work in activity layer height to all of this since before it was on each layer but now all activity layers are combined?
     if (activityOptions.displayMode === 'grouped') {
+      /* Note: here we only pass in a few variables in order to
+        limit the scope of what is reacted to in order to avoid unnecessary re-rendering.
+        A wrapper function is used to provide the other props needed to generate the tree.
+      */
       activityTree = generateActivityTree(
         timeFilteredActivityDirectives,
         timeFilteredSpans,
         activityTreeExpansionMap,
         activityOptions.hierarchyMode,
       );
-      // if (activityOptions.hierarchyMode === 'directive') {
-      //   activityTree = generateActivityTree(timeFilteredActivityDirectives, activityTreeExpansionMap);
-      // } else {
-      //   activityTree = generateActivityTreeFlat(
-      //     timeFilteredActivityDirectives,
-      //     timeFilteredSpans,
-      //     activityTreeExpansionMap,
-      //   );
-      // }
     } else {
       activityTree = [];
     }
-  }
-
-  function getNodeExpanded(id: string, activityTreeExpansionMap: ActivityTreeExpansionMap) {
-    if (!Object.hasOwn(activityTreeExpansionMap, id)) {
-      return false;
-    }
-    return activityTreeExpansionMap[id];
-  }
-
-  function paginate(
-    nodes: ActivityTreeNode[],
-    parentId: string,
-    activityTreeExpansionMap: ActivityTreeExpansionMap,
-    depth = 1,
-  ) {
-    const binSize = 100;
-    if (nodes.length <= binSize) {
-      return nodes;
-    }
-    const newNodes: ActivityTreeNode[] = [];
-    nodes.forEach((node, i) => {
-      const bin = Math.floor(i / binSize);
-      if (!newNodes[bin]) {
-        newNodes[bin] = {
-          children: [],
-          expanded: false,
-          id: '',
-          isLeaf: false,
-          items: [],
-          label: '',
-          type: 'aggregation',
-        };
-      }
-      newNodes[bin].children.push(node);
-      if (node.items) {
-        newNodes[bin].items.push(...node.items);
-      }
-    });
-    newNodes.forEach((node, i) => {
-      const nodeStart = i * binSize ** depth;
-      const nodeEnd = Math.min(nodeStart + node.children.length * depth ** binSize, (i + 1) * binSize ** depth);
-      const label = `[${nodeStart} … ${nodeEnd - 1}]`;
-      node.id = `${parentId}_${label}_page`;
-      node.label = label;
-      node.expanded = getNodeExpanded(node.id, activityTreeExpansionMap);
-    });
-    return paginate(newNodes, parentId, activityTreeExpansionMap, depth + 1);
-  }
-
-  // TODO repeated in LayerActivites, move to a util?
-  function getSpanForActivityDirective(activityDirective: ActivityDirective): Span {
-    const spanId = spanUtilityMaps.directiveIdToSpanIdMap[activityDirective.id];
-    return spansMap[spanId];
   }
 
   function generateActivityTree(
@@ -502,162 +444,19 @@
     spans: Span[],
     activityTreeExpansionMap: ActivityTreeExpansionMap,
     hierarchyMode: ActivityOptions['hierarchyMode'] = 'all',
-  ): ActivityTree {
-    // TODO duplicates appear when you have two layers with the same type
-    const groupedSpans = showSpans && hierarchyMode === 'all' ? groupBy(spans, 'type') : {};
-    const groupedDirectives = showDirectives ? groupBy(directives, 'type') : {};
-    const nodes: ActivityTreeNode[] = [];
-    const allKeys = new Set(Object.keys(groupedSpans).concat(Object.keys(groupedDirectives)));
-    Array.from(allKeys)
-      .sort()
-      .forEach(type => {
-        const spanGroup = groupedSpans[type];
-        const directiveGroup = groupedDirectives[type];
-        const id = type;
-        const expanded = getNodeExpanded(id, activityTreeExpansionMap);
-        const label = type;
-        let children: ActivityTreeNode['children'] = [];
-        let items: ActivityTreeNode['items'] = [];
-        const seenSpans: Record<string, boolean> = {};
-        if (directiveGroup) {
-          directiveGroup.forEach(directive => {
-            let childSpan;
-            if (showSpans) {
-              childSpan = getSpanForActivityDirective(directive);
-              if (childSpan && hierarchyMode === 'all') {
-                seenSpans[childSpan.id] = true;
-              }
-            }
-            if (expanded) {
-              children.push(getDirectiveSubtree(directive, id, activityTreeExpansionMap));
-            }
-            items.push({ directive, ...(childSpan ? { span: childSpan } : null) });
-          });
-        }
-        if (spanGroup && hierarchyMode === 'all') {
-          spanGroup.forEach(span => {
-            if (!seenSpans[span.id]) {
-              if (expanded) {
-                children.push(...getSpanSubtrees(span, id, activityTreeExpansionMap, 'span', filterActivitiesByTime));
-              }
-              items.push({ span });
-            }
-          });
-        }
-        children = paginate(children, id, activityTreeExpansionMap);
-        nodes.push({
-          children,
-          expanded: expanded,
-          id,
-          isLeaf: false,
-          items,
-          label,
-          type: 'aggregation',
-        });
-      });
-    return nodes;
-  }
-
-  function getDirectiveSubtree(
-    directive: ActivityDirective,
-    parentId: string,
-    activityTreeExpansionMap: ActivityTreeExpansionMap,
   ) {
-    let children: ActivityTreeNode[] = [];
-    const id = `${parentId}_${directive.id}`;
-    let span;
-    const expanded = getNodeExpanded(id, activityTreeExpansionMap);
-
-    if (showSpans) {
-      const rootSpanId = spanUtilityMaps.directiveIdToSpanIdMap[directive.id];
-      const rootSpan = spansMap[rootSpanId];
-      if (rootSpan) {
-        span = rootSpan;
-      }
-      if (typeof rootSpanId === 'number') {
-        children = paginate(
-          getSpanSubtrees(rootSpan, id, activityTreeExpansionMap, 'aggregation', filterActivitiesByTime),
-          id,
-          activityTreeExpansionMap,
-        );
-      }
-    }
-
-    return {
-      children,
-      expanded,
-      id,
-      isLeaf: children.length < 1,
-      items: [{ directive, span }],
-      label: directive.name,
-      type: 'directive',
-    } as ActivityTreeNode;
-  }
-
-  function getSpanSubtrees(
-    span: Span,
-    parentId: string,
-    activityTreeExpansionMap: ActivityTreeExpansionMap,
-    type: ActivityTreeNode['type'],
-    filterActivitiesByTime: boolean,
-  ) {
-    let children: ActivityTreeNode[] = [];
-    const spanChildren = spanUtilityMaps.spanIdToChildIdsMap[span.id].map(id => spansMap[id]);
-    if (type === 'aggregation') {
-      // Group by type
-      let computedSpans = spanChildren;
-      if (filterActivitiesByTime) {
-        computedSpans = spanChildren.filter(span => spanInView(span, viewTimeRange));
-      }
-      const groupedSpanChildren = groupBy(computedSpans, 'type');
-      Object.keys(groupedSpanChildren)
-        .sort()
-        .forEach(key => {
-          const spanGroup = groupedSpanChildren[key];
-          const id = `${parentId}_${key}`;
-          const expanded = getNodeExpanded(id, activityTreeExpansionMap);
-          let childrenForKey: ActivityTreeNode[] = [];
-          if (expanded) {
-            spanGroup.forEach(spanChild => {
-              childrenForKey.push(
-                ...getSpanSubtrees(spanChild, id, activityTreeExpansionMap, 'span', filterActivitiesByTime),
-              );
-            });
-            childrenForKey = paginate(childrenForKey, id, activityTreeExpansionMap);
-          }
-          children.push({
-            children: childrenForKey,
-            expanded,
-            id,
-            isLeaf: false,
-            items: spanGroup.map(span => ({ span })),
-            label: key,
-            type: 'aggregation',
-          });
-        });
-    } else if (type === 'span') {
-      const id = `${parentId}_${span.id}`;
-      const expanded = getNodeExpanded(id, activityTreeExpansionMap);
-      const count = spanChildren.length;
-      let childrenForKey: ActivityTreeNode[] = [];
-      if (expanded) {
-        childrenForKey = paginate(
-          getSpanSubtrees(span, id, activityTreeExpansionMap, 'aggregation', filterActivitiesByTime),
-          id,
-          activityTreeExpansionMap,
-        );
-      }
-      children.push({
-        children: childrenForKey,
-        expanded,
-        id,
-        isLeaf: count < 1,
-        items: [{ span }],
-        label: span.type,
-        type: 'span',
-      });
-    }
-    return children;
+    return generateActivityTreeUtil(
+      directives,
+      spans,
+      activityTreeExpansionMap,
+      hierarchyMode,
+      filterActivitiesByTime,
+      spanUtilityMaps,
+      spansMap,
+      showSpans,
+      showDirectives,
+      viewTimeRange,
+    );
   }
 
   function onActivityTreeNodeChange(e: { detail: ActivityTreeNode }) {
